# docs/swagger.yaml
basePath: /
definitions:
  github_com_Its-Ameekh_school_software_backend_internal_apierrors.ErrorResponse:
    properties:
      code:
        description: machine-readable, e.g. "FORBIDDEN"
        type: string
      details:
        description: optional extra context, dev-facing
        type: string
      error:
        type: string
    type: object
...
  github_com_Its-Ameekh_school_software_backend_internal_models.ProgressRemark:
    properties:
      remarks:
        type: string
      studentID:
        type: integer
      term:
        type: string
      updatedAt:
        type: string
      writtenBy:
        type: integer
    type: object
  github_com_Its-Ameekh_school_software_backend_internal_models.ProgressScore:
    properties:
      gradeValue:
        type: string
      gradedBy:
        type: integer
      id:
        type: integer
      maxScore:
        type: number
      scoredValue:
        type: number
      studentID:
        type: integer
      subject:
        type: string
      term:
        type: string
      updatedAt:
        type: string
    type: object
...
  internal_handlers.EnterProgressRequest:
    properties:
      grade_value:
        type: string
      max_score:
        type: number
      remark:
        type: string
      scored_value:
        minimum: 0
        type: number
      student_id:
        type: integer
      subject:
        type: string
      term:
        type: string
    required:
    - grade_value
    - max_score
    - remark
    - scored_value
    - student_id
    - subject
    - term
    type: object
...
  internal_handlers.ProgressViewResponse:
    properties:
      remarks:
        items:
          $ref: '#/definitions/github_com_Its-Ameekh_school_software_backend_internal_models.ProgressRemark'
        type: array
      scores:
        items:
          $ref: '#/definitions/github_com_Its-Ameekh_school_software_backend_internal_models.ProgressScore'
        type: array
      student_id:
        type: integer
    type: object
...
info:
  contact: {}
  description: Backend API for the School Software platform (Preschool Management
    System).
  title: School Software API
  version: "1.0"
paths:
...
  /api/progress/evaluation:
    post:
      consumes:
      - application/json
      description: Upserts student progress scores and global remarks within an isolated
        database transaction.
      parameters:
      - description: Academic evaluation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.EnterProgressRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Saved successfully confirmation statement
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Its-Ameekh_school_software_backend_internal_apierrors.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Its-Ameekh_school_software_backend_internal_apierrors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Enter academic evaluations and remarks
      tags:
      - progress
  /api/progress/view:
    get:
      description: Returns the entire academic scorecard tracking matrix for a student
        ID. Gated for linked child profiles if called by a PARENT role.
      parameters:
      - description: Student ID identifier mapping
        in: query
        name: student_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.ProgressViewResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Its-Ameekh_school_software_backend_internal_apierrors.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Its-Ameekh_school_software_backend_internal_apierrors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: View a student's progress scores and remarks
      tags:
      - progress
...
swagger: "2.0"